+++ Source/Builder.cs
@@ -1,22 +1,27 @@
 
 using System;
-using System.Data.SQLite;
 
 namespace DotSQL.Builder {
-    public class SqliteBuilder: IBuilder {
+    namespace Interfaces {
+        public interface IBuilder {
+            String ConnectionString { get; }
+        }
+    }
+
+    public class SqliteBuilder: Interfaces.IBuilder {
         public String Source { get; set; }
         public String Password { get; set; } = null;
 
         public String ConnectionString {
             get {
-                SQLiteConnectionStringBuilder builder;
+                System.Data.SQLite.SQLiteConnectionStringBuilder builder;
                 if (this.Password == null) {
-                    builder = new SQLiteConnectionStringBuilder {
+                    builder = new System.Data.SQLite.SQLiteConnectionStringBuilder {
                         DataSource = this.Source
                     };
                 }
                 else {
-                    builder = new SQLiteConnectionStringBuilder {
+                    builder = new System.Data.SQLite.SQLiteConnectionStringBuilder {
                         DataSource = this.Source, Password = this.Password
                     };
                 }
@@ -26,7 +31,7 @@ namespace DotSQL.Builder {
         }
     }
 
-    public class MySqlBuilder: IBuilder {
+    public class MySqlBuilder: Interfaces.IBuilder {
         public String Host { get; set; } = "127.0.0.1";
         public Int32 Port { get; set; } = 3306;
         public String UserID { get; set; }
@@ -51,23 +56,8 @@ namespace DotSQL.Builder {
                 return builder.ConnectionString;
             }
         }
-
-        // public MySqlBuilder(String userid, String password, String database, String host = "127.0.0.1", Int32 port = 3306, String charset = "utf8mb4") {
-        //     this.Host = host;
-        //     this.Port = port;
-        //     this.UserID = userid;
-        //     this.Password = password;
-        //     this.Charset = charset;
-        // }
-
-        // public String ToConnectionString() {
-        //     var builder = new MySqlConnector.MySqlConnectionStringBuilder {
-        //         Server = this.Host, Port = (uint)this.Port, UserID = this.UserID, Password = this.Password, Database = this.Database, CharacterSet = this.Charset
-        //     };
-            
-        //     return builder.ConnectionString;
-        // }
     }
 
     public class MariadbBuilder: MySqlBuilder {}
 }
+