+++ Source/Builder/Interfaces.cs
@@ -1,9 +1,0 @@
-
-using System;
-
-
-namespace DotSQL.Builder {
-    public interface IBuilder {
-        String ConnectionString { get; }
-    }
-}