+++ CHANGELOG.md
@@ -6,3 +6,10 @@
 ## 0.0.2
 - change 'DotSQL' of 'DotSQL.SQL' to 'Engine'.
 - update files
+
+<br />
+
+## 0.0.3
+- change 'SQL' namespace to 'Engine'.
+- merge 'Classes.cs' and 'Interfaces.cs' into one file.
+- add 'ExecuteAsync' to Engine.